+++ packages/std/src/results/handle.rs
@@ -7,6 +7,7 @@ use crate::types::Empty;
 
 use super::attribute::Attribute;
 use super::cosmos_msg::CosmosMsg;
+use super::mut_response::MutResponse;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct HandleResponse<T = Empty>
@@ -19,6 +20,15 @@ where
     pub data: Option<Binary>,
 }
 
+impl<T> HandleResponse<T>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    pub fn new() -> Self {
+        Self::default()
+    }
+}
+
 impl<T> Default for HandleResponse<T>
 where
     T: Clone + fmt::Debug + PartialEq + JsonSchema,
@@ -29,5 +39,25 @@ where
             attributes: vec![],
             data: None,
         }
+    }
+}
+
+impl<T> MutResponse<T> for HandleResponse<T>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    fn add_attribute<K: Into<String>, V: Into<String>>(&mut self, key: K, value: V) {
+        self.attributes.push(Attribute {
+            key: key.into(),
+            value: value.into(),
+        });
+    }
+
+    fn add_message<U: Into<CosmosMsg<T>>>(&mut self, msg: U) {
+        self.messages.push(msg.into());
+    }
+
+    fn set_data<U: Into<Binary>>(&mut self, data: U) {
+        self.data = Some(data.into());
     }
 }